% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under the Curve}
\usage{
auc(
  x_vector,
  y_vector,
  x_start = NULL,
  y_start = NULL,
  x_end = NULL,
  y_end = NULL,
  cumulative = FALSE,
  RP_style = FALSE
)
}
\arguments{
\item{x_vector}{vector of the x-coordinates of the executed trajectory}

\item{y_vector}{vector of the y-coordinates of the executed trajectory}

\item{x_start}{x-coordinate of the start point of the ideal trajectory}

\item{y_start}{y-coordinate of the start point of the ideal trajectory}

\item{x_end}{x-coordinate of the end point of the ideal trajectory}

\item{y_end}{y-coordinate of the end point of the ideal trajectory}

\item{cumulative}{whether one single number (cumulative=F)
or an array of cumulative AUCs should be returned}

\item{RP_style}{should areas that stem from a change of direction be added or subtracted?}
}
\value{
AUC as single number or as vector of cumulative AUCs
}
\description{
computes the (cumulative) AUC of a point defined by x and y coordinates,
as compared to an ideal trajectory, as defined by the start and end points.
Importantly, the ideal trajectory is thought of as being of infinite length.
Work via matrix-rotation.
}
\examples{

data("dat_one_trajectory")
auc(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
