% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc2.R
\name{auc2}
\alias{auc2}
\title{Area Under the Curve}
\usage{
auc2(x_vector, y_vector, slope, intercept)
}
\arguments{
\item{x_vector}{vector of the x-coordinates of the executed trajectory}

\item{y_vector}{vector of the y-coordinates of the executed trajectory}

\item{slope}{slope of the ideal trajectory}

\item{intercept}{intercept of the ideal trajectory}
}
\value{
Vector representing the cumulative AUC across course of the trajectory.
}
\description{
This function computes areas under the curve (AUCs) for
trajectory data with the 2D-coordinates x_vector and y_vector and a line,
defined by a slope and an intercept. The arguments x_vector and y_vector
may be scalars or vectors of the same length.
Translation of the script given by
\href{https://doi.org/10.1016/j.cognition.2014.07.012}{
Pfister et al. (2014).}
}
\details{
AUCs are computed by dividing the area in triangular and square-shaped
pieces and adding the up across the trajectory.
}
