% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket-grob.R
\name{bracket-grob}
\alias{bracket-grob}
\alias{makeContent.bracket}
\title{Bracket grob internals}
\usage{
\method{makeContent}{bracket}(x)
}
\description{
Internal grid grob used by \code{geom_bracket()} to draw bracket shapes.
}
\keyword{internal}
