% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_all_but.R
\name{clear_all_but}
\alias{clear_all_but}
\title{Clear the workspace}
\usage{
clear_all_but(keep = NULL)
}
\arguments{
\item{keep}{Names of the items in the workspace to keep. Must be suppled as
strings.}
}
\description{
Clears the workspace but keeps certain items. If 'keep' is not
specified, behaves exactly as the usual 'rm(list = ls())'.
}
\examples{
a = b = c = d = e = 1
print(ls())
clear_all_but(c("a", "b"))
print(ls())

}
