% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{curvature}
\usage{
curvature(x_vector, y_vector)
}
\arguments{
\item{x_vector}{vector of the x-coordinates of the executed trajectory}

\item{y_vector}{vector of the y-coordinates of the executed trajectory}
}
\value{
single number corresponding to the curvature
}
\description{
computes the curvature of a trajectory, defined by arrays
of x and y coordinates, as compared to an ideal trajectory,
as defined by the start and end points of the trajectory
}
\examples{

data("dat_one_trajectory")
curvature(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
