% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_max_acceleration.R
\name{index_max_acceleration}
\alias{index_max_acceleration}
\title{index_max_acceleration}
\usage{
index_max_acceleration(x_vector, y_vector, absolute = FALSE)
}
\arguments{
\item{x_vector}{vector of the x-coordinates of the executed trajectory}

\item{y_vector}{vector of the y-coordinates of the executed trajectory}

\item{absolute}{should negative accelerations (i.e., deceleration) be included?}
}
\value{
single number indicating the number of peak acceleration
}
\description{
computes the index of the peak acceleration of a trajectory,
defined by arrays of x and y coordinates, and assumed to be equidistant in time
}
\examples{

data("dat_one_trajectory")
index_max_acceleration(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
