% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_max_dist.R
\name{index_max_dist}
\alias{index_max_dist}
\title{index_max_dist}
\usage{
index_max_dist(x_vector, y_vector)
}
\arguments{
\item{x_vector}{vector of the x-coordinates of the executed trajectory}

\item{y_vector}{vector of the y-coordinates of the executed trajectory}
}
\value{
single number indicating the index of peak velocity
}
\description{
computes the index of the peak velocity of a trajectory,
defined by arrays of x and y coordinates, and assumed to be equidistant in time
}
\examples{

data("dat_one_trajectory")
index_max_dist(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
