% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_events.R
\name{make_events}
\alias{make_events}
\title{A Function to Generate Precipitation Events}
\usage{
make_events(df = precip_data, rain_col_name = "rain_cm",
  dateTime_col_name = "dateTime", mit_hours = 6, ID = "ID")
}
\arguments{
\item{df}{Name of data frame with precipitation data.}

\item{rain_col_name}{Name of column with rainfall data. Defaults to "rain_cm".}

\item{dateTime_col_name}{Name of the column with timestamps. Defaults to "dateTime".}

\item{mit_hours}{The time in hours to use as the minimum inter-event time. Defaults to 6.0.}

\item{ID}{an identifying word that will be used to create identifiers for events in combination with the mit number}
}
\description{
This function allows you to isolate events from a record of rainfall.
}
\examples{
cat_function()
}
\keyword{cats}

