% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ad.R
\name{max_ad}
\alias{max_ad}
\title{Maximum Absolute Deviation}
\usage{
max_ad(
  x_vector,
  y_vector,
  x_start = NULL,
  y_start = NULL,
  x_end = NULL,
  y_end = NULL
)
}
\arguments{
\item{x_vector}{vector of the x-coordinates of the executed trajectory}

\item{y_vector}{vector of the y-coordinates of the executed trajectory}

\item{x_start}{x-coordinate of the start point of the ideal trajectory}

\item{y_start}{y-coordinate of the start point of the ideal trajectory}

\item{x_end}{x-coordinate of the end point of the ideal trajectory}

\item{y_end}{y-coordinate of the end point of the ideal trajectory}
}
\value{
MAD as single number. Can be negative or positive.
}
\description{
computes the maximum absolute deviation of a vector of points
defined by x and y coordinates,
as compared to an ideal trajectory, as defined by the start and end points.
Importantly, the ideal trajectory is thought of as being of infinite length.
Works via matrix-rotation.
}
\examples{

data("dat_one_trajectory")
max_ad(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
