% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files_from_dir.R
\name{merge_files_from_dir}
\alias{merge_files_from_dir}
\title{Merges Files of the Same Format}
\usage{
merge_files_from_dir(
  regex_exp,
  dir_path = NULL,
  debug = "no",
  include_file_information = FALSE,
  ...
)
}
\arguments{
\item{regex_exp}{A regular expression.
Only file names which match the regular expression will be returned.}

\item{dir_path}{a character vector of full path names;
the default corresponds to the working directory, \code{\link[base]{getwd}}.
Tilde expansion (see \code{\link[base]{path.expand}}) is performed.
Missing values will be ignored.
Elements with a marked encoding will be converted to the native encoding
(and if that fails, considered non-existent).
See \code{\link[base]{list.files}}.}

\item{debug}{debug options. Must be one of "no", "return.names", or
"print.bad.files". Defaults to "no".}

\item{include_file_information}{whether to include some path information to
the returned data set.}

\item{...}{additional arguments that will be supplied to the underlying
\code{\link[data.table]{fread}} function}
}
\value{
merged dataset
}
\description{
crawls the supplied `dir_path` and searches for files matching
`regex_exp`. It then reads in those files and merges the data into one data
frame via appending. Hence, files must have the same data format
(number and identity of columns).
}
