% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_crosses.R
\name{point_crosses}
\alias{point_crosses}
\title{point_crosses}
\usage{
point_crosses(numeric_array, relevant_point = 0)
}
\arguments{
\item{numeric_array}{array of numbers ordered by their time of appearance}

\item{relevant_point}{number which has to be crossed}
}
\value{
number of times that numeric_array crosses the relevant_point
}
\description{
checks how often a certain point is being crossed by an number sequence
where the order of the array indicates timeadjacency
}
\examples{

data("dat_one_trajectory")
point_crosses(dat_one_trajectory$xvals)

}
