% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampen.R
\name{sampen}
\alias{sampen}
\title{sampen}
\usage{
sampen(
  timeseries_array,
  dimensions = 5,
  tolerance = 0.2,
  standardise = TRUE,
  tidy = FALSE
)
}
\arguments{
\item{timeseries_array}{array of numbers over which the sampen is to be computed}

\item{dimensions}{number of embedding dimensions for which to compute the sampen
Sometimes also called "template length"}

\item{tolerance}{the tolerance for the comparisons of two number sequences}

\item{standardise}{whether the tolerance is to be understood as absolute
values or as standardized values}

\item{tidy}{whether the output of the function should be RP-style
or a single number so it can be applied with tidyverse's "mutate"}
}
\value{
returns an array of length "dimensions" with the respective
sampEns or a single number, the last element of this array
}
\description{
computes the sample Entropy, as based on a Matlab script of Roland Pfister.
comes with ABSOLUTELY NO WARRANTY, as this is solely a translation!
}
\examples{

data("dat_one_trajectory")
sampen(dat_one_trajectory$xvals)

}
