% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-errorbar-grob.R
\name{segmentErrorbar-grob}
\alias{segmentErrorbar-grob}
\alias{makeContent.segmentErrorbar}
\title{Segment errorbar grob internals}
\usage{
\method{makeContent}{segmentErrorbar}(x)
}
\description{
Internal grid grob used by \code{geom_segment_errorbar()} to draw error-bars
}
\keyword{internal}
