% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc2.R
\name{auc2}
\alias{auc2}
\title{Area Under the Curve}
\usage{
auc2(x_vector, y_vector, slope, intercept)
}
\arguments{
\item{x_vector}{Vector of the x-coordinates of the executed trajectory.}

\item{y_vector}{Vector of the y-coordinates of the executed trajectory.}

\item{slope}{Slope of the ideal trajectory.}

\item{intercept}{Intercept of the ideal trajectory.}
}
\value{
Vector with the cumulative AUC across course of the trajectory.
}
\description{
Computes areas under the curve (AUCs) for
trajectory data with the 2D-coordinates x_vector and y_vector and a line,
defined by a slope and an intercept. The arguments x_vector and y_vector
may be scalars or vectors of the same length.
Translation of the script given by
Pfister et al. (2014) <\doi{10.1016/j.cognition.2014.07.012}>.
}
\details{
AUCs are computed by dividing the area in triangular and
square-shaped pieces and adding the up across the trajectory.
}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
