% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_all_but.R
\name{clear_all_but}
\alias{clear_all_but}
\title{Clear the global workspace}
\usage{
clear_all_but(keep = NULL)
}
\arguments{
\item{keep}{Names of the items in the workspace to keep. Must be supplied as
vector of strings.}
}
\description{
Clears the global workspace but keeps certain items.
   If \code{keep} is not specified, it behaves exactly as the usual
   \code{\link[base:rm]{rm(list = ls())}}.
}
\examples{
## declare variables
a = b = c = d = e = 1
print(ls())

## remove some of the variables
clear_all_but(c("a", "b"))
print(ls())

}
