% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampen.R
\name{sampen}
\alias{sampen}
\title{SAMPle ENtropy}
\usage{
sampen(
  timeseries_array,
  dimensions = 2,
  tolerance = 0.2,
  standardize = TRUE,
  use_diff = FALSE
)
}
\arguments{
\item{timeseries_array}{array of numbers over which the sampEn is to be computed}

\item{dimensions}{number of embedding dimensions for which to compute the sampEn.
Sometimes also called "template length"}

\item{tolerance}{the tolerance for the comparisons of two number sequences}

\item{standardize}{whether to standardize the timeseries_array}

\item{use_diff}{whether to use the differences between adjacent points}
}
\value{
a single number, the sample entropy for the given parameters
}
\description{
computes the sample Entropy, as given by Richman & Moorman (2000),
which can be found at https://doi.org/10.1152/ajpheart.2000.278.6.H2039
}
\examples{

data("dat_one_trajectory")
sampen(dat_one_trajectory$xvals)

}
