% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampen2.R
\name{sampen2}
\alias{sampen2}
\title{SAMPle ENtropy}
\usage{
sampen2(timeseries_array, dimensions = 5, tolerance = 0.2, standardize = FALSE)
}
\arguments{
\item{timeseries_array}{input signal vector}

\item{dimensions}{maximum template length (default M=5).
Sometimes also called "embedding dimensions"}

\item{tolerance}{matching threshold (default r=.2)}

\item{standardize}{whether the tolerance is to be understood as absolute
values or as standardized values}
}
\value{
returns an array of length "dimensions" with the respective sampEns
(0,1,..,dimensions-1)
}
\description{
computes the sample Entropy, as based on a Matlab script of Roland Pfister.
The matlab script can be found at https://osf.io/am6yp/
and the corresponding publications at https://doi.org/10.3758/s13428-020-01409-0
and https://doi.org/10.1016/j.cognition.2014.07.012
}
\examples{

data("dat_one_trajectory")
sampen2(dat_one_trajectory$xvals)

}
