% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampen2.R
\name{sampen2}
\alias{sampen2}
\title{sampen2}
\usage{
sampen2(
  timeseries_array,
  dimensions = 2,
  tolerance = 0.2 * sd(timeseries_array),
  standardise = FALSE
)
}
\arguments{
\item{timeseries_array}{array of numbers over which the sampEn is to be computed}

\item{dimensions}{number of embedding dimensions for which to compute the sampEn.
Sometimes also called "template length"}

\item{tolerance}{the tolerance for the comparisons of two number sequences}

\item{standardise}{whether to standardize the timeseries_array}
}
\value{
a single number, the sample Entropy for the given parameters
}
\description{
computes the sample Entropy, as I understood it
comes with ABSOLUTELY NO WARRANTY, not certain whether it is completely correct
}
\examples{

data("dat_one_trajectory")
sampen2(dat_one_trajectory$xvals)

}
